upvotes

+++ RevatureConnect/src/main/java/com/revature/controllers/ResponseController.java
@@ -28,7 +28,7 @@ public class ResponseController {
     private ResponseService rservice; 
 
     @GetMapping("/responses/{postId}")
-    public ResponseEntity<List<Response>> getresponsesbypostid(@PathVariable("postId") long postId) { 
+    public ResponseEntity<List<Response>> getresponsesbypostid(@PathVariable("postId") Integer postId) { 
         List<Response> rs = rservice.getResponsesByPostId(postId);
         return new ResponseEntity<List<Response>>(rs, HttpStatus.OK); 
     }
@@ -53,13 +53,13 @@ public class ResponseController {
     }
     
     @DeleteMapping("/responses/delete/{id}")
-    public ResponseEntity<Response> delete(@PathVariable("id") Long id) {
+    public ResponseEntity<Response> delete(@PathVariable("id") Integer id) {
         rservice.deleteResponse(id);
         return new ResponseEntity<Response>(HttpStatus.OK);
     }
     
     @GetMapping("/response/{id}")
-    public ResponseEntity<Response> get(@PathVariable("id") long id) {
+    public ResponseEntity<Response> get(@PathVariable("id") Integer id) {
     	Response response = rservice.getResponseById(id);
     	return new ResponseEntity<Response> (response, HttpStatus.OK);
     }
